+++ ChangeLog.md
@@ -1,12 +1,13 @@
 # Change Log #
 
 ## 0.35 ##
-_8 April 2021_
+_12 April 2021_
 - Bug fix to correctly generate a stereo track for non-4K movies with conditions. Swapped `if [ ${str05VideoHeight}="2160" ]` for `if [ ${str05VideoHeight} -eq 2160 ]`
 - Swapped over to new IP ranges from NAS mounts
 - Change `--target 576p=1500` to `--target 1500` for DVDs as this height isn't supported by `other-transcode`
 - Add in an option for 480p updated targets
 - Changed the methodology to extract forced subtitle checks from using the stream title ("Forced") to using the forced flag setting (`str05SubtitleForcedPresence` from `.disposition.forced`)
+- Added a new audio option to `--add-audio <Index>` for occasional usage whilst also maintaining mono|stereo|surround space (overriding the default stereo in `other-transcode`)